+++ src/main/java/com/getpcpanel/commands/command/Command.java
@@ -10,7 +10,7 @@ import lombok.extern.log4j.Log4j2;
 @SuppressWarnings("InstanceofThis")
 @JsonTypeInfo(use = JsonTypeInfo.Id.CLASS, property = "_type")
 public abstract class Command {
-    public Runnable toRunnable(String deviceId, Integer vol) {
+    public Runnable toRunnable(boolean initial, String deviceId, Integer vol) {
         if (this instanceof ButtonAction ba) {
             return ba.toRunnable();
         }
@@ -18,7 +18,7 @@ public abstract class Command {
             return da.toRunnable(deviceId);
         }
         if (vol != null && this instanceof DialAction da) {
-            return da.toRunnable(vol);
+            return da.toRunnable(initial, vol);
         }
         log.error("Unable to convert {} to Runnable ({}, {})", this, deviceId, vol);
         return () -> {