+++ src/main/java/com/getpcpanel/commands/command/CommandVoiceMeeterBasic.java
@@ -23,7 +23,7 @@ public class CommandVoiceMeeterBasic extends CommandVoiceMeeter implements DialA
     }
 
     @Override
-    public void execute(int level) {
+    public void execute(boolean initial, int level) {
         var voiceMeeter = MainFX.getBean(Voicemeeter.class);
         if (voiceMeeter.login()) {
             voiceMeeter.controlLevel(ct, index, dt, level);