+++ src/main/java/com/getpcpanel/commands/command/CommandVolumeFocus.java
@@ -7,7 +7,7 @@ import lombok.ToString;
 @ToString(callSuper = true)
 public class CommandVolumeFocus extends CommandVolume implements DialAction {
     @Override
-    public void execute(int volume) {
+    public void execute(boolean initial, int volume) {
         getSndCtrl().setFocusVolume(volume / 100f);
     }
 }