+++ src/main/java/com/getpcpanel/commands/command/CommandVolumeProcess.java
@@ -1,9 +1,11 @@
 package com.getpcpanel.commands.command;
 
+import java.util.HashSet;
 import java.util.List;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
+import com.getpcpanel.cpp.MuteType;
 
 import lombok.Getter;
 import lombok.ToString;
@@ -13,16 +15,21 @@ import lombok.ToString;
 public class CommandVolumeProcess extends CommandVolume implements DialAction {
     private final List<String> processName;
     private final String device;
+    private final boolean unMuteOnVolumeChange;
 
     @JsonCreator
-    public CommandVolumeProcess(@JsonProperty("processName") List<String> processName, @JsonProperty("device") String device) {
+    public CommandVolumeProcess(@JsonProperty("processName") List<String> processName, @JsonProperty("device") String device, @JsonProperty("isUnMuteOnVolumeChange") boolean unMuteOnVolumeChange) {
         this.processName = processName;
         this.device = device;
+        this.unMuteOnVolumeChange = unMuteOnVolumeChange;
     }
 
     @Override
-    public void execute(int volume) {
+    public void execute(boolean initial, int volume) {
         var snd = getSndCtrl();
+        if (!initial && unMuteOnVolumeChange) {
+            snd.muteProcesses(new HashSet<>(processName), MuteType.unmute);
+        }
         processName.forEach(process -> snd.setProcessVolume(process, device, volume / 100f));
     }
 }