+++ src/main/java/com/getpcpanel/ui/PickProcessesController.java
@@ -18,6 +18,7 @@ import javafx.scene.layout.Priority;
 import javafx.scene.layout.VBox;
 import javafx.stage.Stage;
 import lombok.RequiredArgsConstructor;
+import lombok.Setter;
 import one.util.streamex.StreamEx;
 
 @Component
@@ -31,7 +32,7 @@ public class PickProcessesController {
     private final FxHelper fxHelper;
 
     @FXML private VBox pickRows;
-    private PickType pickType;
+    @Setter private PickType pickType;
 
     public void initialize() {
         ensureLastEmpty();
@@ -41,8 +42,7 @@ public class PickProcessesController {
         return StreamEx.of(pickRows.getChildren()).select(Pane.class).map(pane -> pane.getChildren().get(0)).select(TextField.class).map(TextField::getText).filter(StringUtils::isNotBlank).toList();
     }
 
-    public void setSelection(PickType type, Collection<String> volData) {
-        pickType = type;
+    public void setSelection(Collection<String> volData) {
         StreamEx.of(volData).map(StringUtils::trimToNull).nonNull().map(this::createProcessRow).forEach(pickRows.getChildren()::add);
         pickRows.getChildren().add(createProcessRow(""));
         removeEmptyIfNotLast();